make dynamic route checking more stable
