+++ src/libs/Navigation/helpers/getPathFromState.ts
@@ -1,26 +1,12 @@
 import {findFocusedRoute, getPathFromState as RNGetPathFromState} from '@react-navigation/native';
 import type {NavigationState, PartialState} from '@react-navigation/routers';
 import {config, normalizedConfigs} from '@libs/Navigation/linkingConfig/config';
-import type {DynamicRouteSuffix} from '@src/ROUTES';
 import type {Screen} from '@src/SCREENS';
 import getDynamicRouteQueryParams from './dynamicRoutesUtils/getDynamicRouteQueryParams';
-import {dynamicRoutePaths} from './dynamicRoutesUtils/isDynamicRouteSuffix';
+import isDynamicRouteScreen from './dynamicRoutesUtils/isDynamicRouteScreen';
 import splitPathAndQuery from './dynamicRoutesUtils/splitPathAndQuery';
 
 type State = NavigationState | Omit<PartialState<NavigationState>, 'stale'>;
-
-/**
- * Checks if a screen name is a dynamic route screen
- */
-function isDynamicRouteScreen(screenName: Screen): boolean {
-    const screenPath = normalizedConfigs[screenName]?.path;
-
-    if (!screenPath) {
-        return false;
-    }
-
-    return dynamicRoutePaths.has(screenPath as DynamicRouteSuffix);
-}
 
 /**
  * Resolves a single path segment: if it's a `:param` placeholder, replaces it
